refactor: move cName function to separate file
